Lil edits and all

+++ styletransfer/options/base_options.py
@@ -25,7 +25,7 @@ class BaseOptions():
         parser.add_argument('--checkpoints_dir', type=str, default='./checkpoints', help='models are saved here')
         # model parameters
         parser.add_argument('--reader', type=str, default='librosa', help='audio reader')
-        parser.add_argument('--duration', type=int, default=8, help='duration of audio')
+        parser.add_argument('--duration', type=float, default=8, help='duration of audio')
         parser.add_argument('--model', type=str, default='cycle_gan', help='chooses which model to use. [cycle_gan | pix2pix | test | colorization]')
         parser.add_argument('--discriminator', type=str, default='conv1d', help='architecture of discriminator')
         parser.add_argument('--transformer', type=str, default='none', help='specify generator transformer architecture')
